+++ modules/metabase/versions.tf
@@ -1,10 +1,11 @@
 terraform {
-  required_version = ">= 1.0"
+  required_version = ">= 1.14.3"
 
   required_providers {
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = "~> 2.30.0"
+      version = "~> 3.0.1"
+
     }
   }
 }