+++ modules/rustfs/versions.tf
@@ -1,14 +1,15 @@
 terraform {
-  required_version = ">= 1.0"
+  required_version = ">= 1.14.3"
 
   required_providers {
     helm = {
       source  = "hashicorp/helm"
-      version = "~> 2.16"
+      version = "~> 3.1.1"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = "~> 2.0"
+      version = "~> 3.0.1"
+
     }
   }
 }